pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "django-cors-headers>=4.9.0",
    "whitenoise>=6.12.0",
    # "mysqlclient>=2.2.8",
    # "psycopg2-binary>=2.9.11",
    "gunicorn>=25.1.0",
    "django-ninja>=1.6.0",
    "django-unfold>=0.85.0",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["config"]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]
